# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent API",
    "version": "1.0.0"
  },
...
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/prompts": {
      "get": {
        "operationId": "listPrompts",
        "parameters": [],
        "tags": ["prompts"],
        "description": "List all saved prompts",
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PromptRow"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Response for 400",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/invalid-json"],
                              "type": "string"
                            },
                            "detail": {
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["The request body is not valid JSON"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/validation-failed"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": ["The input data is invalid"],
                              "type": "string"
                            },
                            "errors": {
                              "type": "object",
                              "additionalProperties": {}
                            }
                          },
                          "required": ["code", "detail", "errors"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Validation failed"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          },
          "415": {
            "description": "Response for 415",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": [
                                "GET requests cannot have a content type header"
                              ],
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": ["Missing content type header"],
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createPrompt",
        "parameters": [],
        "tags": ["prompts"],
        "description": "Create a new prompt",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["content", "name"],
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "minLength": 1
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response for 201",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptRow"
                }
              }
            }
          },
          "400": {
            "description": "Response for 400",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/invalid-json"],
                              "type": "string"
                            },
                            "detail": {
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["The request body is not valid JSON"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/validation-failed"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": ["The input data is invalid"],
                              "type": "string"
                            },
                            "errors": {
                              "type": "object",
                              "additionalProperties": {}
                            }
                          },
                          "required": ["code", "detail", "errors"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Validation failed"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["message"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          },
          "415": {
            "description": "Response for 415",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": [
                                "GET requests cannot have a content type header"
                              ],
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "enum": ["Missing content type header"],
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "cause": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "enum": ["api/unsupported-media-type"],
                              "type": "string"
                            },
                            "detail": {
                              "type": "string"
                            }
                          },
                          "required": ["code", "detail"],
                          "additionalProperties": false
                        },
                        "message": {
                          "enum": ["Unsupported Media Type"],
                          "type": "string"
                        }
                      },
                      "required": ["cause", "message"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PromptRow": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "number"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": ["content", "created_at", "id", "name", "version"],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Use Authorization: Bearer <token> for session tokens or API keys."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Use X-API-Key: <key>."
      }
    }
  }
}